backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic==2.5.0
google-auth==2.23.0
python-jose[cryptography]==3.3.0
